Update 4
Added button component that takes in children and props and displays children.  Added error component.  Added apollolink library that links network and application layers and handles separate errors between each layer.  Added errormessage component to the profile component for errorhandling if there is a problem with the gql query or data returned.  Added Mutation component to the Repository item component.  The mutation component is an apollo-react component that composes the mutation query and takes a predefined gql mutation and variables, performs the mutation on the repo and displays the result.  The mutation component checks if the viewer has starred the repository and if not displays a button component that performs the add star mutation on click.  The button also displays the current total stargazer count.