Update 6
Added gql fragment file for re-using parts of the query that are used many times.  Added updateadd/removeStar function that takes in the client and repository data then writes the updated mutation fragment to the cache with the totalcount incremented or decremented 1.  Added the fragmentupdates as an update parameter in the mutation function within the Mutation component.