Update 7
Added watch/unwatch gql mutation.  Added updateSubscription mutation component that contains a button. The mutation is made on button click.  The updateSubscription data is read as a fragment.  A fragment query is made to the apollo cache, the watcher count is updated and current subscription status is changed to its opposite.  Then the fragment is written back into the cache and the ui updated.