update 8
Updated main query to include cursor variable and get page info from first query result.  The page info returns the endcursor position and whether there is more data to be returned.  Added optimistic response to the add/remove star and watch mutations.  Optimistic response will update the data to the expected state before the data returns and if it returns not as expected throws an error.  Added fragment component to the repositoryList component that handles loading state and also updates the rest of the data.