Update 14
Added prefetched query/caching to issues on hover.  When the user hovers over the issues buttons the query is made before the user clicks and the data cached.  This is another example of imparative querying and saves the user loading time.  Added ApolloConsumer component to handle getting the apollo client data for the react app.  Extracted the onChangeissues button into a filterIssues component to handle getting and displaying the issues.  Added prefetchIssues function to check if current issue state is none and if it is perform the query and store the data.